[threaded-animations] many `css-view-transitions` WPT tests crash with "Threaded Time-based Animations" enabled https://bugs.webkit.org/show_bug.cgi?id=304077
Reviewed by Simon Fraser.

Fix the faulty logic when assembling keyframe values for `filter` and `backdrop-filter`
in `AcceleratedEffect::validateFilters()`. Indeed, if only a single keyframe was explicitly
provided, we would fail to reach the `else` statement for the case where processing a keyframe
that was not explicitly affecting the provided property.

We now take a simpler approach by keeping a boolean flag that indicates whether we've encountered
an explicit "to" keyframe and, if that flag is not set, add the implicit value.

Note that there is no test change in this patch since the flag is not yet enabled on bots. This
was caught in preparation of that running animation tests locally using
`--experimental-feature ThreadedTimeBasedAnimationsEnabled=true`.

* Source/WebCore/platform/animation/AcceleratedEffect.cpp:
(WebCore::AcceleratedEffect::validateFilters):

Canonical link: https://commits.webkit.org/304377@main