[appearance: base] Create a child pseudo render element for the ::checkmark pseudo style https://bugs.webkit.org/show_bug.cgi?id=303606 rdar://165893004
Reviewed by Antti Koivisto.

A child checkmark box will be created for ::checkmark pseudo style. It will be
created only if the appearance is `base` and the ::checkmark is not styled with
`display: none;` or `display: contents;`.

This checkmark box will be removed from the parent renderer when the appearance is
not `base` or ::checkmark is styled with `display: none;` or `display: contents;`.

The UA style sheet will have a rule to set ::checkmark style with `display: none;`
when it is unchecked. For now this rule will be added in the test style sheet.

Tests: fast/forms/appearance-base/appearance-base-checkable-checkmark-dynamic.html
       fast/forms/appearance-base/appearance-base-checkable-checkmark-initial.html

* LayoutTests/fast/forms/appearance-base/appearance-base-checkable-checkmark-dynamic-expected.html: Added.
* LayoutTests/fast/forms/appearance-base/appearance-base-checkable-checkmark-dynamic.html: Added.
* LayoutTests/fast/forms/appearance-base/appearance-base-checkable-checkmark-initial-expected.html: Added.
* LayoutTests/fast/forms/appearance-base/appearance-base-checkable-checkmark-initial.html: Added.
* Source/WebCore/html/HTMLInputElement.cpp:
(WebCore::HTMLInputElement::setChecked):
* Source/WebCore/rendering/updating/RenderTreeBuilder.cpp:
(WebCore::RenderTreeBuilder::updateAfterDescendants):
* Source/WebCore/rendering/updating/RenderTreeBuilderFormControls.cpp:
(WebCore::RenderTreeBuilder::FormControls::updateCheckmark):
(WebCore::RenderTreeBuilder::FormControls::updateAfterDescendants):
* Source/WebCore/rendering/updating/RenderTreeBuilderFormControls.h:

Canonical link: https://commits.webkit.org/304390@main