ASAN_ILL | WebCore::LocalFrameViewLayoutContext::performLayout; WebCore::LocalFrameViewLayoutContext::layout; WebCore::Document::updateLayout https://bugs.webkit.org/show_bug.cgi?id=298078 rdar://157023263
Reviewed by Alan Baradlay.

The placeholder text in the RenderTextControl is treated as excluded content and uses special
layout logic. This special layout logic does not consider that the placeholder can be absolutely
positioned like in the testcase, resulting in a render tree that ends up being dirty.

To prevent this problem, do not allow position values other than the default 'static', which
actually matches the specification [1], since it states a limited number of properties apply
to the ::placeholder pseudo-element, which does not include the position property.

[1] https://drafts.csswg.org/css-pseudo/#placeholder-pseudo

* LayoutTests/fast/forms/textarea-with-absolute-placeholder-crash-expected.txt: Added.
* LayoutTests/fast/forms/textarea-with-absolute-placeholder-crash.html: Added.
* Source/WebCore/css/html.css:
(::placeholder):

Originally-landed-as: 297297.10@webkit-2025.7-embargoed (b4e1ccfd7e27). rdar://166339811
Canonical link: https://commits.webkit.org/304398@main