ASAN_ILL | Layout::LineBuilder::initialize; Layout::LineBuilder::layoutInlineContent; Layout::IntrinsicWidthHandler::computedIntrinsicWidthForConstrain https://bugs.webkit.org/show_bug.cgi?id=298301 rdar://154646612
Reviewed by Alan Baradlay.

(RenderObjects relying on a) RenderMarquee can run preferred width
computation but not run layout between tree mutations. In general,
we skip invalidation for removal of a never-laid-out renderer. This
patch drops this optimization if the removed renderer is in a
subtree of a never-laid-out RenderMarquee.

* LayoutTests/fast/layout/removing-descendant-from-a-never-laid-out-marquee-crash-expected.txt: Added.
* LayoutTests/fast/layout/removing-descendant-from-a-never-laid-out-marquee-crash.html: Added.
* Source/WebCore/html/HTMLMarqueeElement.h:
* Source/WebCore/rendering/updating/RenderTreeBuilder.cpp:
(WebCore::isRenderMarquee): New helper to check if a RenderObject has an associated RenderMarquee.
(WebCore::isWithinNeverLaidOutRenderMarqueeSubtree): New helper to check if a RenderObject is withing a never laid out subtree of a node with an associated RenderMarquee.
(WebCore::invalidateLineLayout): For invalidation if we are in a never laid out RenderMarquee subtree.

Originally-landed-as: 297297.8@webkit-2025.7-embargoed (e78d76a4e51b). rdar://166340697
Canonical link: https://commits.webkit.org/304418@main