[threaded-animations] tests with animated motion path properties crash with "Threaded Time-based Animations" enabled https://bugs.webkit.org/show_bug.cgi?id=304112 rdar://166444990
Reviewed by Alan Baradlay.

Several animation tests animating Motion Path properties crash in debug builds due to failing
`ASSERT(canBlend(from, to))` assertions in `AcceleratedEffectOffsetAnchor` and other structs
of that type. This is due to not passing in a renderer to the `AcceleratedEffectValues` constructor
when creating the `AcceleratedEffect` keyframes. As a result we deal with empty values when trying
to blend in the remote layer tree.

Known affected tests are:

- webanimations/accelerated-animations-and-motion-path.html
- webanimations/offset-distance-animation-yields-compositing.html
- webanimations/offset-rotate-animation-yields-compositing.html

Note that there is no test change in this patch since the flag is not yet enabled on bots. This
was caught in preparation of that running animation tests locally using
`--experimental-feature ThreadedTimeBasedAnimationsEnabled=true`.

* Source/WebCore/platform/animation/AcceleratedEffect.cpp:
(WebCore::AcceleratedEffect::AcceleratedEffect):

Canonical link: https://commits.webkit.org/304425@main