[Xcode] Changes to IDL file list in DerivedSources.make can cause incremental build failures https://bugs.webkit.org/show_bug.cgi?id=304096 rdar://166421164
Reviewed by Simon Fraser.

IDLFileNamesList.txt is derived from the JS_BINDING_IDLS list in the
Makefile, so the text listing needs to be regenerated whenever the
Makefile changes. Otherwise, removing an IDL will not regenerate the
list, which will cause the generator to fail.

To avoid spuriously re-running the generator, change
IDLFileNamesList.txt to be an order-only prerequisite of the generator
rules. This means that the generator requires that the file exists,
but will not be invalidated whenever the file updates. That's file; what
invalidates the generator is changes to IDL content itself.

* Source/WebCore/DerivedSources.make:

Canonical link: https://commits.webkit.org/304430@main