AX: Inform assistive technologies that WebKit performs it's own text stitching https://bugs.webkit.org/show_bug.cgi?id=304071 rdar://166390461
Reviewed by Joshua Hoffman.

Return AXPerformsOwnTextStitching in accessibilityAttributeNames to inform assistive technologies that they do not need
to perform text stitching because WebKit does it for them. accessibilityAttributeNames was chosen because the key AT
in question (VoiceOver) already requests and caches this for every element, so adding this new attribute there is as
low of a performance cost as can be achieved.

* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper accessibilityAttributeNames]):

Canonical link: https://commits.webkit.org/304469@main