AX: Accessibility text stitching doesn't happen within links, causing unnecessary accessibility tree fragmentation https://bugs.webkit.org/show_bug.cgi?id=304187 rdar://166544999
Reviewed by Joshua Hoffman.

Prior to this commit, we didn't stitch any text at all inside links, which is a bad user experience. The intention was
to break up the stitch when entering a link, but the logic was simply incorrect, so this commit fixes it.

Test: accessibility/text-stitching-inside-links.html

* LayoutTests/accessibility/text-stitching-inside-links-expected.txt: Added.
* LayoutTests/accessibility/text-stitching-inside-links.html: Added.
* LayoutTests/platform/glib/TestExpectations: Skip new test.
* LayoutTests/platform/ios/TestExpectations: Mark new test as passing.
* Source/WebCore/accessibility/AXStitchUtilities.cpp:
(WebCore::isStitchBreakingElement):
(WebCore::shouldStopStitchingAt):
* Source/WebCore/accessibility/AXStitchUtilities.h:
* Source/WebCore/accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::stitchGroups const):

Canonical link: https://commits.webkit.org/304491@main