Revert "(286869@main): <details> disclosure triangle is now tiny with `system-ui`" https://bugs.webkit.org/show_bug.cgi?id=304208 rdar://164545409
Reviewed by Brent Fulgham and Tim Nguyen.

This reverts commit 4593e7d6b65aa56d69fad5312d902a4ba12e5c38.

At https://commits.webkit.org/280433@main [1] we made the rendering of the
disclosure triangle symbol consistent for all directions (writing-mode, rtl,
ltr). This meant rendering the right small triangle variation according to
style, which is compatible with other engines that also use such variants.

However, a bug was filed about the new glyphs being too small. We then
adapted the changes from [1] to use the bigger variant at
https://commits.webkit.org/299530@main [2].

We didn't notice that for some fonts the bigger variant can be rendered with
emoji representation on iOS, which is the bug reported here.

We tried different approaches:

1. Using the text variation selector (U+FE0E) along with the bigger triangle.
This would force the selection of text representation. However, this leads to
inconsistent rendering. The glyphs chosen for different directions have
different sizes, which is not desired.

2. Using a fixed symbol and rotating it during painting. This would introduce
a difference between layout and painting. Additionally, this works for list
markers but fails for ::before content that uses counter(), requiring
implementation for each renderer type.

Given the drawbacks of the 2 approaches listed, we have decided to go back to
the small triangle variant without the text variation selector and therefore
revert [2]. This, as mentioned, is compatible with other engines.

Canonical link: https://commits.webkit.org/304525@main