Playback of application/ogg blob media is broken rdar://163119790 https://bugs.webkit.org/show_bug.cgi?id=301509
Reviewed by Jean-Yves Avenard.

When loading through blob, WebCoreAVFResourceLoader is used.
`application/ogg` mime type does not lead to a valid UTI and AV player is not able to recover itself.

Mime Sniffer is used but it will still return `application/ogg`.
We update MediaPlayerPrivateAVFoundationObjC::createAVAssetForURL to use `audio/ogg` in case `application/ogg` is the preferred mime type.
We only do this for blob URLs and should follow-up for all URLs.
We skip the test on iOS due to the lack of ogg support there.

Test: http/wpt/media/blob-and-ogg.html
Canonical link: https://commits.webkit.org/304530@main