[Accessibility] HAVE_ADDITIONAL_NS_ACCESSIBILITY_CONSTANTS's SDK check causes availability issues in back-deployment https://bugs.webkit.org/show_bug.cgi?id=303936
Reviewed by Tyler Wilcock and Richard Robinson.

In https://commits.webkit.org/303588@main, the conditions for
HAVE_ADDITIONAL_NS_ACCESSIBILITY_CONSTANTS changed to check SDK version
(__MAC_OS_X_VERSION_MAX_ALLOWED_MAX_ALLOWED) instead of deployment
target version (__MAC_OS_X_VERSION_MIN_REQUIRED). This is wrong, because
the constants in AppKit are symbols, so they require runtime support.
i.e. we can only bind to the AppKit symbols when we are deploying to an
OS that will have them, and we must redeclare the constants we need
otherwise. Change it back.

This revealed issues relating to include order:
CocoaAccessibilityConstants.h has to be included *after* AppKit, because
on back-deployed builds it can define macros that clash with AppKit's
declarations. Include order sensitivity is fine in our builds, but not
in modules. Yet modules are self-contained and supposed to be
commutative to include order.

What's actually happening is that WebCore's module verifier invocation
has a bad flag (-fno-implicit-module-maps) that was preventing it from
actually building modules. Removing the flag reveals new problems with
the modularization, so, disable the verifier for now until we can fix
them.

* Source/WTF/wtf/PlatformHave.h:
* Source/WebCore/Configurations/WebCore.xcconfig:
* Tools/DumpRenderTree/mac/AccessibilityUIElementMac.mm: Reorder
* Tools/TestWebKitAPI/Tests/WebKit Swift/Foundation+Extras.swift: As a
  drive-by, add an @available annotation to make the tests build when
  back-deploying to macOS 14.6.

Canonical link: https://commits.webkit.org/304546@main