Introduce MemoryDump printing utility for dataLog() https://bugs.webkit.org/show_bug.cgi?id=304214 rdar://166570876
Reviewed by Dan Hecht and Keith Miller.

It's sometimes helpful in trace output to print relatively large chunks of raw data such as
stack frames in the traditional hex dump form. This patch introduces a new printing
plug-in for dataLog() to do just that.

Examples:

    dataLogLn("Interesting data:", MemoryDump(startPtr, endPtr));
    dataLogLn("Interesting data:", MemoryDump(std::span(data)));

The style checker is unhappy with the usage of 'out.printf()' in the code added to
PrintStream.cpp, however that is consistent with the rest of that file.

Tests: Tools/TestWebKitAPI/Tests/WTF/MemoryDump.cpp
Canonical link: https://commits.webkit.org/304560@main