[Swift in WebKit] Correct the availability of WebGPU in some configurations https://bugs.webkit.org/show_bug.cgi?id=304281 rdar://166646850
Reviewed by Mike Wyrzykowski.

- The enablement of `ENABLE_WEBGPU_SWIFT` for macOS was conditionalized on the architecture.
However, in some parts of the build, this value is `undefined`. This leads to incredibly
confusing and surprising results. To address this, and to allow module verification to
actually be conditionalized on `ENABLE_WEBGPU_SWIFT`, just enable `ENABLE_WEBGPU_SWIFT` at
build time on all architectures on macOS.

- Disable `ENABLE_WEBGPU_SWIFT` on macOS SDKs that do not have a modularized WTF library, since
that is an invalid combination.

* Source/WebGPU/Configurations/WebGPU.xcconfig:
* Source/WebGPU/WebGPU/CommandEncoder.swift:
(WebGPU.finish(_:)):

Canonical link: https://commits.webkit.org/304571@main