Big5 decoder doesn't recover to emit ASCII after an invalid leading byte https://bugs.webkit.org/show_bug.cgi?id=304238 rdar://166672674
Reviewed by Anne van Kesteren.

* LayoutTests/imported/w3c/web-platform-tests/encoding/textdecoder-eof.any.js:
Added test cases for this bug and a couple other Big5 decoding bugs that were
reported at the same time. This is probably not the ideal place for these tests,
but it's better to have them somewhere inside WPT rather than waiting to find
the perfect place. Anyone is welcome to reorganize and clean up the location of
these in WPT, I will not be offended.

* Source/WebCore/PAL/pal/text/TextCodecCJK.cpp:
(PAL::TextCodecCJK::big5Decode): Added m_prependedByte logic to one place it
was missing.

Canonical link: https://commits.webkit.org/304591@main