REGRESSION(304573@main): [GTK][WPE] API test expectations don't work anymore because the default webkitpy port name has changed https://bugs.webkit.org/show_bug.cgi?id=304332
Unreviewed partial revert fix.

304573@main changed the webkitpy port name from gtk and wpe to gtk-wk2
and wpe-wk2 and that has caused that the API Test Expectations don't
match anymore.

This recovers back the previous webkitpy names for GTK and WPE because this
port names are used on several parts of the webkitpy code and not only on
the API test expectation file.

If there is really a desire to change the default webkitpy port name in this
case, then a more careful approach and some discussion is needed.

* Tools/Scripts/webkitpy/common/test_expectations_unittest.py:
* Tools/Scripts/webkitpy/port/glib.py:
(GLibPort):
(GLibPort.determine_full_port_name):

Canonical link: https://commits.webkit.org/304594@main