[WPE] Link jsc and other jsc-related binaries (testb3, testair, TestJavaScriptCore, etc) against libWPEWebKit https://bugs.webkit.org/show_bug.cgi?id=304326
Reviewed by Nikolas Zimmermann.

We are having issues with the size of the Debug builds on WPE, currently at ~5GB are causing
issues with the upload of the build products on the CI.

An important part of this problem is the fact that in WPE there is no shared libjavascriptcore
library like in GTK, and instead of linking against the shared library libWPEWebKit we are
embedding (statically linking) all the libjavascriptcore code into the binaries. So we end with
9 binaries around 700MB in Debug when we could shrink all those binaries to a few megabytes by
just linking dynamically against libWPEWebKit.

This patch makes all those jsc-related binaries to link against libWPEWebKit. The net effect is
a reduction in size from 5GB to 3.3GB for the debug.zip built-product.

* Source/JavaScriptCore/shell/PlatformWPE.cmake: Added.
* Tools/TestWebKitAPI/PlatformWPE.cmake:

Canonical link: https://commits.webkit.org/304666@main